removed tools; moved SegmentMaker and ScoreTemplate to content directory
