+++ setup.py
@@ -4,11 +4,7 @@ if __name__ == "__main__":
     setup(
         author="Tsz Kiu Pang",
         author_email="osamupang@gmail.com",
-        install_requires=[
-            "numpy",
-            "abjad",
-            "abjad-ext-nauert",
-        ],
+        install_requires=["numpy", "abjad", "abjad-ext-nauert",],
         name="utsu",
         packages=["utsu"],
         version="0.1",
