+++ utsu/ScoreTemplate.py
@@ -13,15 +13,10 @@ class ScoreTemplate(abjad.ScoreTemplate):
             simultaneous=True,
         )
         piano_LH_staff = abjad.Staff(
-            [abjad.Voice(name="Piano LH Voice")],
-            name="LH Staff",
+            [abjad.Voice(name="Piano LH Voice")], name="LH Staff",
         )
         staff_group = abjad.StaffGroup(
-            [piano_RH_staff, piano_LH_staff],
-            name="Piano Staff Group",
+            [piano_RH_staff, piano_LH_staff], name="Piano Staff Group",
         )
-        score = abjad.Score(
-            [staff_group],
-            name="Piano Score",
-        )
+        score = abjad.Score([staff_group], name="Piano Score",)
         return score
