+++ utsu/_lib.py
@@ -14,8 +14,5 @@ sieves = abjad.OrderedDict(
 )
 
 metronome_marks = abjad.OrderedDict(
-    [
-        ("60", abjad.MetronomeMark((1, 4), 60)),
-        ("72", abjad.MetronomeMark((1, 4), 72)),
-    ]
+    [("60", abjad.MetronomeMark((1, 4), 60)), ("72", abjad.MetronomeMark((1, 4), 72)),]
 )
